after setting the trigger app, send the requisite CONNECT/DISCONNECT Intents
When the user changes the panic app config, then the current app needs to
send Intents to the previous app, and the currently configed app to let
them know about the changes.  This is done by sending an ACTION_DISCONNECT
Intent to the previous app, and an ACTION_CONNECT Intent to the newly
configured app.